requirements.txt
python-telegram-bot[job-queue]>=20.0,<23.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
google-auth>=2.22.0
openai>=1.0.0
